refactor(ontology-coded-action-deploy): one script per action, not one CLI per file
solution_release.py was a 532-line, 7-subcommand, 22-function argparse tree: a
mini-service, not a utility. It becomes eight scripts named for what they do,
one per action the skill's phases actually invoke, over three shared modules.

  scaffold_solution.py   (was solution_scaffold.py)   phase 1
  stage_jobs.py                                        phase 2
  next_version.py                                      phase 3
  build_package.py                                     phase 3
  publish_package.py                                   phase 3   MUTATES
  deploy_release.py                                    phase 3   MUTATES
  resolve_folder_id.py                                 phase 4
  await_release.py                                     phase 4
  patch_action_ttl.py    (was ttl_patch.py)            phase 5   MUTATES

  _uip.py        the CLI boundary: argv lists only, stdout for data, stderr for errors
  _solution.py   reading solution state, on disk and as the tenant has it
  _staging.py    the staging tree and the derived manifest

Split by cohesion rather than by verb count. `pack` stops being a public verb
because nothing but `publish` ever called it, and the stage/build/publish chain
lives in _staging.py so the three CLIs are thin rather than three copies.
folder_id moved into _solution.py once it turned out two scripts need it.

This structurally fixes a defect the review found: solution_release.py resolved
`solution_name()` before dispatching, so `folder-id` and `await` exited 1
without SOLUTION_SRC despite their own docs saying otherwise. Per-action scripts
do not call it at all, and both now run without it. Verified.

Every script answers --describe with its own contract as JSON: name, purpose,
phase, inputs, outputs, whether it mutates, exit codes. Handled before argparse
by a shared `described()` helper, so a script with required positionals can
still be interrogated. That is the determinism-through-script half; the skill
supplies sequence and context, and the next commit adds the table that refers
to them.

Verified offline end to end: template scaffold, then stage, produces main.ts at
the project root with a manifest byte-identical to the committed Studio Web
golden. All nine scripts answer --help and --describe. The 30 validator tests
are unaffected.
